Make package names to case-insensitive
* (Required for Gradle support)
* Make all other packages lowercase
* Moved com.GlehuDonsa to com.glehudonsa.battlearena
